feat(scrape): add basic poster cropping and config files
The scrape sub-command now pulls and save the cover of the title scraped. It detects which format the images are and crops them as needed.
For now, the poster cropping is very basic, but a preset system will be added, pulling presets from the configuration file added by this commit.