feat(scrape): add scraper result amalgamation
A config file (in YAML) is now provided with sensible defaults. It defines the priority of the scrapers for the various fields.
During scraping, results from all the scrapers are collected inside a dictionnary and the final results are then built using the priorities defined in the config file.

Fixes #8